Fixed variable misspelling between classes
RendererWorker inherits from RendererModernGL (and currently can only inherit from it). The code comments define it as
```# same logic as parent class but with sleep when queue is full```
There's a variable counting how many frames this worker has generated so far. Both ```__render__``` and ```check_consistency``` mistakenly call the variable ```self.frames_generated``` instead of ```self.n_frames_generated```, as the parent class does.

Because ```self.update_uniforms``` isn't overridden and uses the parent's variable name, the n_frame in ```update_uniform_time``` is always 0 (or if temporal_sampling_mode is a list, the first element). this causes both list- and continuous-generations to always generate the same images. Random (and therefore ```generate_data_single_shader.py```) just so happens to avoid this by not caring which frame number we are currently generating.

This commit fixes the variable name.