QoL cross-platform change
https://docs.python.org/3/library/platform.html
The code currently assumes by default you are running on linux. As far as I can tell, windows uses WGL and OSX uses GLX, neither supporting egl.
This pull request is a slim change to check if you are in linux, and if you aren't use the default backend for your system as determined by modernGL.

I hadn't tested it other than on windows, and an even better solution might be to use the default backend regardless, but I figured it might save someone a headache.
Take it or leave it ¯\_(ツ)_/¯.
Also, fixed a spelling error because it bugged me.